[SERCOM] Regenerated demo for sercom_spi_spi_self_loopback_interrupt_sam_c21n_xpro MH3-27885

+++ apps/sercom/spi/spi_self_loopback_interrupt/firmware/src/config/sam_c21n_xpro/peripheral/sercom/spim/plib_sercom5_spi.c
@@ -164,7 +164,7 @@ bool SERCOM5_SPI_TransferSetup(SPI_TRANSFER_SETUP *setup, uint32_t spiSourceCloc
         if((baudValue > 0) & (baudValue <= 255))
         {
             /* Selection of the Clock Polarity and Clock Phase */
-            SERCOM5_REGS->SPIM.SERCOM_CTRLA |= setup->clockPolarity | setup->clockPhase;
+            SERCOM5_REGS->SPIM.SERCOM_CTRLA |= (uint32_t)setup->clockPolarity | (uint32_t)setup->clockPhase;
 
             /* Selection of the Baud Value */
             SERCOM5_REGS->SPIM.SERCOM_BAUD = baudValue;
